A Link’s Fineness Based on Spatial Relation
The propagation of wireless mobile equipment has given rise to mobile ad hoc networks (MANET), which allow the establishment of networks deprived of infrastructure support. The improvements in these technologies have also resulted in the variety of link quality, scalability, stability and reachability limits. Principally, due to their incapability to guarantee node reliability, MANETs are weak in highly dynamic environment. Since high-speed nodes, low density, and rapid changes in the topology are directly damage networks performances and lifetime. In response, routing is an inspiring issue in these networks. This paper proposed a new routing scheme in MANETS. This technique is integrated in the optimized link state routing (OLSR). In the projected scheme, the author defines a link fineness based on spatial relation between mobile nodes, which exploits different parameters, speed, direction and acceleration, including the sent and received of hello packets. Correspondingly, the proposed selected multipoint relays (MPRs) based the projected methods is included in the algorithm. The scheme elected more stable MPRs with high link’s quality, more neighborhood degree and higher willingness as MPR that can increase the reachability and stability of the network even in an extremely dynamic situation. This version of OLSR is verified by the network simulator (NS3) under the Radom Waypoint mobility model. Results revealed an important improvement for SLF_OLSR. In addition, this scheme can be integrated into other protocols in MANETs.